backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Euro Hotel Authentication Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "motor>=3.3.0",
    "beanie>=1.23.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.3",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",
    "razorpay>=1.3.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
]
